+++ src/de/bielefeld/umweltamt/aui/mappings/elka_sync/EStandort.java
@@ -20,6 +20,7 @@
  */
 package de.bielefeld.umweltamt.aui.mappings.elka_sync;
 
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
@@ -35,8 +36,8 @@ import de.bielefeld.umweltamt.aui.mappings.DatabaseQuery;
  */
 public class EStandort implements java.io.Serializable {
 
-	private Integer nr;
-	private Integer origNr;
+	private BigInteger nr;
+	private BigInteger origNr;
 	private EAdresse adresse;
 	private String gemeindeId;
 	private String gemarkung;
@@ -56,11 +57,11 @@ public class EStandort implements java.io.Serializable {
 	public EStandort() {
 	}
 
-	public EStandort(Integer nr) {
+	public EStandort(BigInteger nr) {
 		this.nr = nr;
 	}
 
-	public EStandort(Integer nr, Integer origNr, EAdresse adresse,
+	public EStandort(BigInteger nr, BigInteger origNr, EAdresse adresse,
 			String gemeindeId, String gemarkung,
 			String flur, String flurstuecke, Float e32, Float n32,
 			Boolean industrieabwasserTog, Boolean niederschlagswasserTog,
@@ -85,21 +86,21 @@ public class EStandort implements java.io.Serializable {
 		this.herkunft = herkunft;
 	}
 
-	public Integer getNr() {
+	public BigInteger getNr() {
 		return this.nr;
 	}
 
-	public void setNr(Integer nr) {
+	public void setNr(BigInteger nr) {
 		this.nr = nr;
 	}
 
 	@JsonIgnore
-	public Integer getOrigNr() {
+	public BigInteger getOrigNr() {
 		return this.origNr;
 	}
 	
 	@JsonIgnore
-	public void setOrigNr(Integer origNr) {
+	public void setOrigNr(BigInteger origNr) {
 		this.origNr =  origNr;
 	}
 	
@@ -253,7 +254,7 @@ public class EStandort implements java.io.Serializable {
      *         if one exists,
      *         <code>null</code> otherwise
      */
-    public static EStandort findById(java.lang.Integer id) {
+    public static EStandort findById(BigInteger id) {
         return (EStandort)
             new DatabaseAccess().get(EStandort.class, id);
     }
